resolved merge conflict with new queue handlers to update state
